refactor: ReservationTimeRequestDto 역할분리
- ReservationTime 전환을 위한 역할만 하도록 연관 코드 수정
- Reservation 에 연관된 코드는 TimeDto 를 생성해 적용
- 관련 테스트 코드 수정
