+++ src/main/java/roomescape/reservationtime/dto/ReservationTimeRequestDto.java
@@ -4,36 +4,17 @@ import jakarta.validation.constraints.NotBlank;
 
 public class ReservationTimeRequestDto {
 
-    private Long timeId;
     @NotBlank(message = "예약 시간을 입력해주세요")
     private String startAt;
 
-    public ReservationTimeRequestDto(Long timeId) {
-        this.timeId = timeId;
+    public ReservationTimeRequestDto() {
     }
 
     public ReservationTimeRequestDto(String startAt) {
         this.startAt = startAt;
     }
 
-    public ReservationTimeRequestDto(Long timeId, String startAt) {
-        this.timeId = timeId;
-        this.startAt = startAt;
-    }
-
-    public Long getTimeId() {
-        return timeId;
-    }
-
     public String getStartAt() {
         return startAt;
-    }
-
-    @Override
-    public String toString() {
-        return "{ " +
-                "timeId=" + timeId +
-                ", startAt='" + startAt + '\'' +
-                '}';
     }
 }
