+++ src/test/java/roomescape/globalfixture/dto/ReservationDtoFixture.java
@@ -1,6 +1,7 @@
 package roomescape.globalfixture.dto;
 
 import roomescape.reservation.dto.ReservationRequestDto;
+import roomescape.reservation.dto.TimeDto;
 import roomescape.reservationtheme.dto.ReservationThemeRequestDto;
 import roomescape.reservationtime.dto.ReservationTimeRequestDto;
 
@@ -8,7 +9,7 @@ public class ReservationDtoFixture {
 
     public static ReservationRequestDto createReservationDto(){
         final ReservationThemeRequestDto reservationThemeDto = ReservationThemeDtoFixture.createReservationThemeDto();
-        final ReservationTimeRequestDto reservationTimeRequestDto = ReservationTimeDtoFixture.createReservationTimeRequestDto();
-        return new ReservationRequestDto("김준성", "2025-08-15", reservationTimeRequestDto, reservationThemeDto);
+        final TimeDto timeDto = TimeDtoFixture.timeDtoCreate();
+        return new ReservationRequestDto(null, "2025-08-15", timeDto, reservationThemeDto);
     }
 }
