+++ src/test/java/roomescape/globalfixture/dto/ReservationTimeDtoFixture.java
@@ -5,6 +5,6 @@ import roomescape.reservationtime.dto.ReservationTimeRequestDto;
 public class ReservationTimeDtoFixture {
 
     public static ReservationTimeRequestDto createReservationTimeRequestDto(){
-        return new ReservationTimeRequestDto(1L, "10:00");
+        return new ReservationTimeRequestDto("10:00");
     }
 }
