+++ src/test/java/roomescape/globalfixture/dto/TimeDtoFixture.java
@@ -1,0 +1,10 @@
+package roomescape.globalfixture.dto;
+
+import roomescape.reservation.dto.TimeDto;
+
+public class TimeDtoFixture {
+
+    public static TimeDto timeDtoCreate(){
+        return new TimeDto(1L, "10:00");
+    }
+}
